feat: refactor directory structure of workspace management
